+++ itachallenge-challenge/src/main/java/com/itachallenge/common/exception/GlobalExceptionHandler.java
@@ -192,4 +192,9 @@ public class GlobalExceptionHandler {
         return ResponseEntity.status(HttpStatus.CONFLICT).body(new MessageDto(ex.getMessage()));
     }
 
+    @ExceptionHandler(ForbiddenException.class)
+    public ResponseEntity<MessageDto> handleForbiddenException(ForbiddenException ex) {
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new MessageDto(ex.getMessage()));
+    }
+
 }